+++ core/src/commonMain/kotlin/NavigationNode.kt
@@ -12,11 +12,12 @@ import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 
-abstract class NavigationNode<Config : Base, Base> {
+abstract class NavigationNode<Config : Base, Base>(
+    open val id: NavigationNodeId = NavigationNodeId()
+) {
     protected open val log: KSLog by lazy {
         TagLogger(toString())
     }
-    open val id: NavigationNodeId = NavigationNodeId()
 
     abstract val chain: NavigationChain<Base>
     abstract val configState: StateFlow<Config>
@@ -84,8 +85,8 @@ abstract class NavigationNode<Config : Base, Base> {
         log.d { "onDestroy" }
     }
 
-    fun createEmptySubChain(): NavigationChain<Base> {
-        return NavigationChain<Base>(this, chain.nodeFactory).also {
+    fun createEmptySubChain(id: NavigationChainId? = null): NavigationChain<Base> {
+        return NavigationChain<Base>(this, chain.nodeFactory, id).also {
             _subchainsFlow.value += it
         }
     }
@@ -139,7 +140,11 @@ abstract class NavigationNode<Config : Base, Base> {
         return subscope.coroutineContext.job
     }
 
-    class Empty<T>(override val chain: NavigationChain<T>, config: T) : NavigationNode<T, T>() {
+    class Empty<T>(
+        override val chain: NavigationChain<T>,
+        config: T,
+        id: NavigationNodeId = NavigationNodeId()
+    ) : NavigationNode<T, T>(id) {
         override val configState: StateFlow<T> = MutableStateFlow(config).asStateFlow()
     }
 }