+++ core/src/commonMain/kotlin/extensions/Chain.kt
@@ -1,12 +1,13 @@
 package dev.inmo.navigation.core.extensions
 
 import dev.inmo.micro_utils.common.Diff
-import dev.inmo.micro_utils.common.applyDiff
 import dev.inmo.micro_utils.common.diff
 import dev.inmo.navigation.core.NavigationChain
+import dev.inmo.navigation.core.NavigationChainId
 import dev.inmo.navigation.core.NavigationNode
+import dev.inmo.navigation.core.NavigationNodeId
+import dev.inmo.navigation.core.chainOrNodeEither
 import kotlinx.coroutines.flow.*
-
 
 val <Base> NavigationChain<Base>.onNodesStackDiffFlow: Flow<Diff<NavigationNode<out Base, Base>>>
     get() = flow {
@@ -25,3 +26,92 @@ val <Base> NavigationChain<Base>.onNodeReplacedFlow
     get() = onNodesStackDiffFlow.map { it.replaced }.filter { it.isNotEmpty() }
 
 fun <Base> NavigationChain<Base>.rootChain(): NavigationChain<Base> = parentNode ?.chain ?.rootChain() ?: this
+
+fun <Base> NavigationChain<Base>.findNode(id: NavigationNodeId): NavigationNode<*, Base>? = stackFlow.value.firstNotNullOfOrNull {
+    it.findNode(id)
+}
+
+fun <Base> NavigationChain<Base>.findChain(id: NavigationChainId): NavigationChain<Base>? = if (this.id == id) {
+    this
+} else {
+    stackFlow.value.firstNotNullOfOrNull { it.findChain(id) }
+}
+
+
+// Drop/replace/push by node id
+
+/**
+ * Shortcut for method [dev.inmo.navigation.core.ChainOrNodeEither].[dropInSubTree]
+ */
+fun <Base> NavigationChain<Base>.dropInSubTree(
+    id: NavigationNodeId
+): Boolean = chainOrNodeEither().dropInSubTree(id)
+
+/**
+ * Shortcut for method [dev.inmo.navigation.core.ChainOrNodeEither].[dropNodeInSubTree]
+ */
+fun <Base> NavigationChain<Base>.dropNodeInSubTree(id: String) = chainOrNodeEither().dropNodeInSubTree(id)
+
+/**
+ * Shortcut for method [dev.inmo.navigation.core.ChainOrNodeEither].[replaceInSubTree]
+ */
+fun <Base> NavigationChain<Base>.replaceInSubTree(
+    id: NavigationNodeId,
+    config: Base,
+): Boolean = chainOrNodeEither().replaceInSubTree(id, config)
+
+/**
+ * Shortcut for method [dev.inmo.navigation.core.ChainOrNodeEither].[replaceInSubTree]
+ */
+fun <Base> NavigationChain<Base>.replaceInSubTree(
+    id: String,
+    config: Base
+) = chainOrNodeEither().replaceInSubTree(id, config)
+
+/**
+ * Shortcut for method [dev.inmo.navigation.core.ChainOrNodeEither].[pushInSubTree]
+ */
+fun <Base> NavigationChain<Base>.pushInSubTree(
+    inChainWith: NavigationNodeId,
+    config: Base
+): Boolean = chainOrNodeEither().pushInSubTree(inChainWith, config)
+
+/**
+ * Shortcut for method [dev.inmo.navigation.core.ChainOrNodeEither].[pushInSubTreeByNodeId]
+ */
+fun <Base> NavigationChain<Base>.pushInSubTreeByNodeId(
+    inChainWithNodeId: String,
+    config: Base
+) = chainOrNodeEither().pushInSubTreeByNodeId(inChainWithNodeId, config)
+
+// Drop/push by chain id
+
+/**
+ * Shortcut for method [dev.inmo.navigation.core.ChainOrNodeEither].[dropInSubTree]
+ */
+fun <Base> NavigationChain<Base>.dropInSubTree(
+    id: NavigationChainId
+) = chainOrNodeEither().dropInSubTree(id)
+
+/**
+ * Shortcut for method [dev.inmo.navigation.core.ChainOrNodeEither].[dropChainInSubTree]
+ */
+fun <Base> NavigationChain<Base>.dropChainInSubTree(
+    id: String
+) = chainOrNodeEither().dropChainInSubTree(id)
+
+/**
+ * Shortcut for method [dev.inmo.navigation.core.ChainOrNodeEither].[pushInSubTree]
+ */
+fun <Base> NavigationChain<Base>.pushInSubTree(
+    inChainWithNodeId: NavigationChainId,
+    config: Base
+) = chainOrNodeEither().pushInSubTree(inChainWithNodeId, config)
+
+/**
+ * Shortcut for method [dev.inmo.navigation.core.ChainOrNodeEither].[pushInSubTreeByChainId]
+ */
+fun <Base> NavigationChain<Base>.pushInSubTreeByChainId(
+    inChainWithNodeId: String,
+    config: Base
+) = chainOrNodeEither().pushInSubTreeByChainId(inChainWithNodeId, config)