+++ core/src/commonMain/kotlin/visiter/NavigationNodeVisitingCallback.kt
@@ -1,0 +1,5 @@
+package dev.inmo.navigation.core.visiter
+
+import dev.inmo.navigation.core.NavigationNode
+
+typealias NavigationNodeVisitingCallback<Base> = (NavigationNode<out Base, Base>) -> Unit